# src/main/resources/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: API
  license:
    name: MIT
paths:
  /clients:
    get:
      summary: List all clients
      operationId: listClients
      responses:
        '200':
          description: An array of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clients"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a client
      operationId: createClients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Client"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /clients/{clientId}:
    get:
      summary: Info for a specific client
      operationId: showClientById
      parameters:
        - name: clientId
          in: path
          required: true
          description: The id of the client to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /clients/import:
    post:
      summary: Import Clients CSV data
      operationId: importClients
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Accepted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /clients/top:
    get:
      summary: List top clients
      operationId: listTopClients
      parameters:
        - name: limit
          in: query
          required: true
          description: The limit of top client to retrieve
          schema:
            type: integer
        - name: from
          in: query
          required: true
          description: Date from which clients loyalty points are taken into consideration
          schema:
            type: string
            format: date
      responses:
        '200':
          description: An array of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clients"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /appointments:
    get:
      summary: List all appointments
      operationId: listAppointments
      responses:
        '200':
          description: An array of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointments"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create an appointment
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Appointment"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /appointments/{appointmentId}:
    get:
      summary: Info for a specific appointment
      operationId: showAppointmentById
      parameters:
        - name: appointmentId
          in: path
          required: true
          description: The id of the appointment to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /appointments/import:
    post:
      summary: Import Appointments CSV data
      operationId: importAppointments
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Accepted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products:
    get:
      summary: List all products
      operationId: listProducts
      responses:
        '200':
          description: An array of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Products"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products/{productId}:
    get:
      summary: Info for a specific product
      operationId: showProductById
      parameters:
        - name: productId
          in: path
          required: true
          description: The id of the product to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a specific product
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          required: true
          description: The id of the product to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a specific product
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          required: true
          description: The id of the product to delete
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Expected response to a valid request
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products/importServices:
    post:
      summary: Import Services CSV data
      operationId: importServices
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Accepted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products/importPurchases:
    post:
      summary: Import Purchases CSV data
      operationId: importPurchases
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Accepted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Client:
      type: object
      required:
        - firstName
        - lastName
        - email
        - phone
        - gender
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        gender:
          $ref: "#/components/schemas/Gender"
        banned:
          type: boolean
    Clients:
      type: array
      items:
        $ref: "#/components/schemas/Client"
    Gender:
      enum:
        - Male
        - Female
    Appointment:
      type: object
      required:
        - clientId
        - startTime
        - endTime
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    Appointments:
      type: array
      items:
        $ref: "#/components/schemas/Appointment"
    Product:
      type: object
      required:
        - appointmentId
        - name
        - price
        - loyaltyPoints
        - type
      properties:
        id:
          type: string
          format: uuid
        appointmentId:
          type: string
          format: uuid
        name:
          type: string
        price:
          type: number
          format: double
        loyaltyPoints:
          type: integer
        productType:
          $ref: "#/components/schemas/ProductType"
    Products:
      type: array
      items:
        $ref: "#/components/schemas/Product"
    ProductType:
      enum:
        - service
        - purchase
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string